pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "confluence-agent"
version = "0.1.0"
description = "Multi-source → Local LLM → Confluence Knowledge Agent"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "trafilatura>=1.8",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
pdf = ["pymupdf>=1.24"]
docx = ["python-docx>=1.0"]
youtube = ["yt-dlp>=2024.1"]
excel = ["pandas>=2.0", "openpyxl>=3.1"]
all = [
    "pymupdf>=1.24",
    "python-docx>=1.0",
    "yt-dlp>=2024.1",
    "pandas>=2.0",
    "openpyxl>=3.1",
]

[project.scripts]
confluence-agent = "src.main:main"
